requirements.txt
tronpy==0.5.0
Django==5.1.3
django-background-tasks==1.2.8
python-dotenv==1.0.1